combine into helmUninstall as per review
