+++ vars/helmUninstall.groovy
@@ -1,8 +1,24 @@
 import uk.gov.hmcts.contino.DockerImage
 import uk.gov.hmcts.contino.Kubectl
 import uk.gov.hmcts.contino.Helm
+import uk.gov.hmcts.contino.PipelineCallbacksRunner
 
-def call(DockerImage dockerImage, Map params) {
+def call(DockerImage dockerImage, Map params, PipelineCallbacksRunner pcr) {
+
+  try {
+    stageWithAgent("Uninstall Helm Release - ${params.environment}", params.product) {
+      pcr.callAround("helmReleaseUninstall:${params.environment}") {
+        withAksClient(params.subscription, params.environment, params.product) {
+          uninstallRelease(dockerImage, params)
+        }
+      }
+    }
+  } catch (ignored) {
+    echo "Unable to uninstall this helm release."
+  }
+}
+
+def uninstallRelease(DockerImage dockerImage, Map params) {
   
   def subscription = params.subscription
